Number Wizard Code
Udemy Number Wizard Program for use in Unity. C#.

The goal of this small project was to create a program that could home in on a number you chose between 1 and 1000.

After implementing the base code, I expanded on the program by allowing the user to select their own number range between 1 and 5000 and having the computer generate a random number that it would then carry into the Update function.

The next step for this project is to create some basic UI functionality. As of right now, the program just prints to the console.